Improvised error printing.
Lexer attaches position of the token along with the token so that it
can help in error printing. This way the TOK_CONSUMED global veriable is
rendered useless, hence removed. Changed the error printing format. Now
using sprintf so that the sub program can write error msgs wihtout being
held back by the positions and use of format identifiers.