Added basic semantic checking
It recognizes multiple definitions of NON-TERMINALs and also checks if
all the declared NON_TERMINALs are defined or not. It checks all this
and throws an error as well using the error printing struct and
function. A vector for the ast_node pointers have been setup to help in
future optimizations like left-factoring.